Freecad/STL files for AMS7262 adafruit module
These are the design files for the Adafruit AMS7262 color module holder. These sit on top of PVC pipes that are cut to keep the face of the module to within 1/4 of an inch of the rotating disk bottom. The tolerance is very tight so be careful to make the bottom of the platter as flat as possible. The cover is not necessary but may cut down on stray ambient light from confusing the color reading. Note: there is no AMS7262Base1.stl file, just use the 'Base2 file.